Fix issue#32 json-parser: segfault on invalid JSON '{"foo"}'
Parsing a Json string having an invalid object contained lead to a segmentation fault. More in detail, when the object has no property but illegally only a string contained, e.g., '{ "foo" }'. The reason for the bug was that the parser expects an assignment token (':') after reading a string (which is in this context a key). Since there was no further check, the parser consumed all next token (i.e., '{' for object begin, 'foo' as key, '}' was skipped because ':' was assumed, and the next token fetch to read the value lead to a read outside the input string token list, which finally curses a segmentation fault. Fixed this bug by explicitly checking for ':' and the number of available tokens.